STApp/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

pandas==2.0.2
pdf2image==1.17.0
Pillow==9.5.0
PyPDF2==3.0.1
SQLAlchemy==2.0.31
streamlit==1.37.1
